Moving metadata to its own folder
