+++ GameDataParser/Parsers/MapEntityParser.cs
@@ -6,6 +6,7 @@ using System.Text.RegularExpressions;
 using System.Xml;
 using GameDataParser.Crypto.Common;
 using Maple2Storage.Types;
+using Maple2Storage.Types.Metadata;
 using ProtoBuf;
 using System.Configuration;
