+++ Maple2Storage/Types/Metadata/ItemMetadata.cs
@@ -1,7 +1,8 @@
 ﻿using System;
 using System.Xml.Serialization;
 
-namespace Maple2Storage.Types {
+namespace Maple2Storage.Types.Metadata
+{
     [XmlType]
     public class ItemMetadata {
         [XmlElement(Order = 1)]
