+++ MapleServer2/PacketHandlers/Game/MoveFieldHandler.cs
@@ -1,5 +1,5 @@
 ﻿using System.Linq;
-using Maple2Storage.Types;
+using Maple2Storage.Types.Metadata;
 using MaplePacketLib2.Tools;
 using MapleServer2.Constants;
 using MapleServer2.Data.Static;
