+++ MapleServer2/Tools/GameCommandActions.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Maple2Storage.Types;
+using Maple2Storage.Types.Metadata;
 using MapleServer2.Data.Static;
 using MapleServer2.Enums;
 using MapleServer2.Packets;
